McpServerTemplate/pyproject.toml
[project]
name = "McpServerTemplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.6.0",
    # 数据验证和设置管理
    "pydantic>=2.0.0",
    # 日期时间处理 - 最强大的日期库
    "python-dateutil>=2.8.2",
    # URL处理和解析
    "urllib3>=2.0.0",
    # 字符串处理和模板
    "jinja2>=3.1.2",
    # 编码转换工具
    "chardet>=5.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages=["src/McpServerTemplate"]

[project.scripts]
McpServerTemplate = "McpServerTemplate:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "ruff>=0.7.3"
]
